Correct "grad_lipschitz for right-multiplication"
Correct the formula for grad_lipschitz for right-multiplication of a functional with a scalar.

Let f be a functional and g := f * a. The formula currently implemented states that grad_lipschitz(g) = abs(a) * grad_lipschitz(f). However, it is not hard to see that this should be grad_lipschitz(g) = abs(a)**2 * grad_lipschitz(f).